Moves linked resources method to client.
