+++ src/Connector/Client.php
@@ -6,6 +6,7 @@ use Psr\Http\Message\ResponseInterface;
 use GuzzleHttp\Exception\BadResponseException;
 use AcquiaCloudApi\Exception\ApiErrorException;
 use Psr\Http\Message\StreamInterface;
+use AcquiaCloudApi\Exception\LinkedResourceNotImplementedException;
 
 /**
  * Class Client
@@ -199,5 +200,55 @@ class Client implements ClientInterface
     public function addOption($name, $value): void
     {
         $this->options = array_merge_recursive($this->options, [$name => $value]);
+    }
+
+    /**
+     * @param array{type:string, path:string, responseClass:class-string} $link
+     * @return mixed
+     * @throws LinkedResourceNotImplementedException
+     */
+    public function getLinkedResource($link)
+    {
+        // Remove the base URI from the path as this is already added by the Connector when we call request().
+        $path = str_replace(ConnectorInterface::BASE_URI, '', $link['path']);
+        $type = $link['type'];
+        $responseClass = $link['responseClass'];
+
+        $classMap = [
+            'alerts' => '\AcquiaCloudApi\Response\InsightAlertsResponse',
+            'applications' => '\AcquiaCloudApi\Response\ApplicationsResponse',
+            'backups' => '\AcquiaCloudApi\Response\BackupsResponse',
+            'code' => '\AcquiaCloudApi\Response\BranchesResponse',
+            'crons' => '\AcquiaCloudApi\Response\CronsResponse',
+            'databases' => '\AcquiaCloudApi\Response\DatabasesResponse',
+            'domains' => '\AcquiaCloudApi\Response\DomainsResponse',
+            'environments' => '\AcquiaCloudApi\Response\EnvironmentsResponse',
+            'ides' => '\AcquiaCloudApi\Response\IdesResponse',
+            'insight' => '\AcquiaCloudApi\Response\InsightsResponse',
+            'logs' => '\AcquiaCloudApi\Response\LogsResponse',
+            'members' => '\AcquiaCloudApi\Response\MembersResponse',
+            'metrics' => '\AcquiaCloudApi\Response\MetricsResponse',
+            'modules' => '\AcquiaCloudApi\Response\InsightModulesResponse',
+            'notification' => '\AcquiaCloudApi\Response\NotificationResponse',
+            'permissions' => '\AcquiaCloudApi\Response\PermissionsResponse',
+            'self' => $responseClass,
+            'servers' => '\AcquiaCloudApi\Response\ServersResponse',
+            'ssl' => '\AcquiaCloudApi\Response\SslCertificatesResponse',
+            'teams' => '\AcquiaCloudApi\Response\TeamsResponse',
+            'variables' => '\AcquiaCloudApi\Response\VariablesResponse',
+        ];
+
+        // Clear any queries attached to the client to prevent sorts etc being carried through
+        // from the original query.
+        // @TODO this may need to be removed if users wish to filter/sort linked resources.
+        $this->clearQuery();
+
+        if (isset($classMap[$type])) {
+            return new $classMap[$type](
+                $this->request('get', $path)
+            );
+        }
+
+        throw new LinkedResourceNotImplementedException($type . ' link not implemented in this SDK. Please file an issue here: https://github.com/typhonius/acquia-php-sdk-v2/issues');
     }
 }
