+++ src/Connector/ClientInterface.php
@@ -6,6 +6,7 @@ use Psr\Http\Message\ResponseInterface;
 use GuzzleHttp\Exception\BadResponseException;
 use AcquiaCloudApi\Exception\ApiErrorException;
 use Psr\Http\Message\StreamInterface;
+use AcquiaCloudApi\Exception\LinkedResourceNotImplementedException;
 
 /**
  * Interface ClientInterface
@@ -101,4 +102,11 @@ interface ClientInterface
      * @param mixed  $value
      */
     public function addOption($name, $value): void;
+
+    /**
+     * @param array{type:string, path:string, responseClass:class-string} $link
+     * @return mixed
+     * @throws LinkedResourceNotImplementedException
+     */
+    public function getLinkedResource($link);
 }
